requirements.txt
uvicorn~=0.23.1
fastapi~=0.100.1
psycopg2~=2.9.6
pyjwt~=2.8.0
pydantic[email]~=2.1.1
passlib
bcrypt~=4.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19
# mysql==0.0.3
pg8000==1.30.1

